Provide player mark in Play method.
By popular demand, provide the player mark to Play method.

It's possible to infer the mark from looking at the board (since X
always goes first, if there are odd number of free spots on the board
when it's your turn, you must be X, and if there are odd free spots,
you must be O). However, it may be convenient to just be provided that
information alongside with the board. It's easy to do so.

Hopefully this makes the interface a little friendlier to implement.