cmd/tictactoe: split terminal/browser code, add WebAssembly support
Previously, it only worked in a terminal and in browser via GopherJS.
Now it supports both GopherJS and WebAssembly.

Separate the terminal and browser display code. This is needed because
syscall/js package is empty when wasm build constraint isn't specified.
It also makes the code cleaner and removes stdout output in browser.